# Replace the Dictionary with HashSet in AppDomainShutdownMonitor