Benchmark:

|             Method |              argList |     Mean |    Error |   StdDev | Ratio | RatioSD |
|------------------- |--------------------- |---------:|---------:|---------:|------:|--------:|
| 'AddAndRemove Old' | Syste(...)ence] [55] | 679.7 us | 10.42 us |  8.70 us |  1.00 |    0.00 |
|                    |                      |          |          |          |       |         |
| 'AddAndRemove New' |  Syst(...)er]] [113] | 667.7 us | 11.07 us | 10.35 us |     ? |       ? |


|                Method |              argList |     Mean |     Error |    StdDev | Ratio | RatioSD |
|---------------------- |--------------------- |---------:|----------:|----------:|------:|--------:|
| 'AddBeforeRemove Old' | Syste(...)ence] [55] | 1.414 ms | 0.0174 ms | 0.0154 ms |  1.00 |    0.00 |
|                       |                      |          |           |           |       |         |
| 'AddBeforeRemove New' |  Syst(...)er]] [113] | 1.197 ms | 0.0097 ms | 0.0081 ms |     ? |       ? |


|        Method |     Mean |     Error |    StdDev |   Median | Ratio | RatioSD |
|-------------- |---------:|----------:|----------:|---------:|------:|--------:|
| 'AddOnly Old' | 2.886 ms | 0.0574 ms | 0.1160 ms | 2.821 ms |  1.00 |    0.00 |
| 'AddOnly New' | 2.880 ms | 0.0639 ms | 0.1865 ms | 2.827 ms |  1.00 |    0.09 |


Benchmark Code: https://github.com/lindexi/lindexi_gd/tree/b63ba44c2077bcc4136455f16b309fe74a288dcc/AppDomainShutdownMonitor